Mv client to secretly.go; Add package doc comment
